requirements/piplist.txt
absl-py==1.4.0
addict==2.4.0
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.1
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
attributee==0.1.8
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1687884021435/work
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1688813409104/work
chumpy==0.70
click==8.1.6
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
crcmod==1.7
cryptography==41.0.3
cycler==0.11.0
Cython==3.0.0
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
dlib==19.24.99
dotty-dict==1.3.1
dsntnn==0.5.3
exceptiongroup==1.1.2
executing==1.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastjsonschema==2.18.0
fonttools==4.42.0
fqdn==1.5.1
geomloss==0.2.6
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipykernel==6.25.1
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==0.10.0
json-tricks==3.17.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
kiwisolver==1.4.4
lap==0.4.0
lazy_loader==0.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mmcls==0.25.0
mmcv-full==1.6.2
mmdet==2.28.2
mmpose==0.29.0
mmtrack==0.14.0
model-index==0.1.11
motmetrics==1.4.0
munkres==1.1.4
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
notebook==7.0.2
notebook_shim==0.2.3
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.8.0.74
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.17
ordered-set==4.1.0
oss2==2.17.0
overrides==7.4.0
packaging==23.1
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1675487152289/work
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
POT==0.9.2
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pycparser==2.21
pycryptodome==3.18.0
Pygments==2.16.1
pyparsing==3.0.9
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.2
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.9.2
rsa==4.9
scikit-image==0.21.0
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2
terminado==0.17.1
terminaltables==3.1.10
thop==0.1.1.post2209072238
tifffile==2023.7.10
tinycss2==1.2.1
tomli==2.0.1
torch==1.12.0
torch-summary==1.4.5
torchaudio==0.12.0
torchstat==0.0.7
torchvision==0.13.0
tornado==6.3.3
tqdm==4.65.0
traitlets==5.9.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1688315532570/work
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
widgetsnbextension==4.0.8
xmltodict==0.13.0
xtcocotools==1.13
yapf==0.40.1
zipp==3.16.2
